further cleanup of code base added github actions and makefile
