+++ .github/workflows/ci.yml
@@ -1,0 +1,32 @@
+name: Lint & Test
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - '*'
+jobs:
+  build:
+    name: '[test] trade-platform'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@master
+      - uses: actions/setup-go@v1
+        with:
+          go-version: 1.13
+      - name: Download golangci-lint
+        run: curl -sfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh| sh -s -- -b $(go env GOPATH)/bin v1.23.6
+        env:
+          GO111MODULE: 'on'
+      - name: Lint
+        run: |
+          export PATH=$PATH:$(go env GOPATH)/bin
+          golangci-lint run --verbose
+        env:
+          GO111MODULE: 'on'
+      - name: test
+        run: |
+          make test_ci
+        env:
+          GO111MODULE: 'on'
