+++ .gitignore
@@ -26,3 +26,4 @@ _testmain.go
 .DS_Store
 .vscode
 .python-version
+.idea/
