+++ go.mod
@@ -2,4 +2,7 @@ module github.com/thrasher-corp/go-talib
 
 go 1.13
 
-require github.com/stretchr/testify v1.5.1
+require (
+	github.com/stretchr/testify v1.5.1
+	golang.org/dl v0.0.0-20200225205303-b5f8f1376514 // indirect
+)
