+++ indicators/bbands.go
@@ -3,7 +3,6 @@ package indicators
 // BBands - Bollinger Bands
 // upperband, middleband, lowerband = BBands(close, timeperiod=5, nbdevup=2, nbdevdn=2, matype=0)
 func BBands(inReal []float64, inTimePeriod int, inNbDevUp float64, inNbDevDn float64, inMAType MaType) ([]float64, []float64, []float64) {
-
 	outRealUpperBand := make([]float64, len(inReal))
 	outRealMiddleBand := Ma(inReal, inTimePeriod, inMAType)
 	outRealLowerBand := make([]float64, len(inReal))
@@ -11,7 +10,6 @@ func BBands(inReal []float64, inTimePeriod int, inNbDevUp float64, inNbDevDn flo
 	tempBuffer2 := StdDev(inReal, inTimePeriod, 1.0)
 
 	if inNbDevUp == inNbDevDn {
-
 		if inNbDevUp == 1.0 {
 			for i := 0; i < len(inReal); i++ {
 				tempReal := tempBuffer2[i]
