+++ indicators/rsi.go
@@ -2,7 +2,6 @@ package indicators
 
 // Rsi - Relative strength index
 func Rsi(inReal []float64, inTimePeriod int) []float64 {
-
 	outReal := make([]float64, len(inReal))
 
 	if inTimePeriod < 2 {
@@ -35,7 +34,6 @@ func Rsi(inReal []float64, inTimePeriod int) []float64 {
 	prevGain /= float64(inTimePeriod)
 
 	if today > 0 {
-
 		tempValue1 = prevGain + prevLoss
 		if !((-0.00000000000001 < tempValue1) && (tempValue1 < 0.00000000000001)) {
 			outReal[outIdx] = 100.0 * (prevGain / tempValue1)
@@ -43,7 +41,6 @@ func Rsi(inReal []float64, inTimePeriod int) []float64 {
 			outReal[outIdx] = 0.0
 		}
 		outIdx++
-
 	} else {
 
 		for today < 0 {
@@ -64,7 +61,6 @@ func Rsi(inReal []float64, inTimePeriod int) []float64 {
 	}
 
 	for today < len(inReal) {
-
 		tempValue1 = inReal[today]
 		today++
 		tempValue2 = tempValue1 - prevValue
@@ -88,4 +84,23 @@ func Rsi(inReal []float64, inTimePeriod int) []float64 {
 	}
 
 	return outReal
+}
+
+// StochRsi - Stochastic Relative Strength Index
+func StochRsi(inReal []float64, inTimePeriod int, inFastKPeriod int, inFastDPeriod int, inFastDMAType MaType) ([]float64, []float64) {
+	outFastK := make([]float64, len(inReal))
+	outFastD := make([]float64, len(inReal))
+
+	lookbackSTOCHF := (inFastKPeriod - 1) + (inFastDPeriod - 1)
+	lookbackTotal := inTimePeriod + lookbackSTOCHF
+	startIdx := lookbackTotal
+	tempRSIBuffer := Rsi(inReal, inTimePeriod)
+	tempk, tempd := StochF(tempRSIBuffer, tempRSIBuffer, tempRSIBuffer, inFastKPeriod, inFastDPeriod, inFastDMAType)
+
+	for i := startIdx; i < len(inReal); i++ {
+		outFastK[i] = tempk[i]
+		outFastD[i] = tempd[i]
+	}
+
+	return outFastK, outFastD
 }
